+++ src/react/components/autocompleter/XUIAutocompleter.js
@@ -5,8 +5,8 @@ import debounce from 'lodash.debounce';
 import { v4 as uuidv4 } from 'uuid';
 import Picklist from '../picklist/Picklist';
 import XUILoader from '../loader/XUILoader';
-import DropDown from '../dropdown/DropDown';
-import DropDownToggled from '../dropdown/DropDownToggled';
+import XUIDropDown from '../dropdown/XUIDropDown';
+import XUIDropDownToggled from '../dropdown/XUIDropDownToggled';
 import XUITextInput from '../textInput/XUITextInput';
 import { ns } from '../helpers/xuiClassNamespace';
 import { fixedWidthDropdownSizes } from '../dropdown/private/constants';
@@ -327,7 +327,7 @@ export default class XUIAutocompleter extends PureComponent {
     );
 
     const dropdown = (
-      <DropDown
+      <XUIDropDown
         className={dropdownClassName}
         fixedWidth={dropdownFixedWidth}
         footer={footer}
@@ -348,7 +348,7 @@ export default class XUIAutocompleter extends PureComponent {
         ) : (
           children
         )}
-      </DropDown>
+      </XUIDropDown>
     );
 
     const classNames = cn(className, focused && `${ns}-autocompleter--trigger-focus`);
@@ -362,7 +362,7 @@ export default class XUIAutocompleter extends PureComponent {
         onFocus={this.onFocus}
         ref={this.rootNode}
       >
-        <DropDownToggled
+        <XUIDropDownToggled
           ariaRole="combobox"
           closeOnSelect={closeOnSelect}
           closeOnTab={closeOnTab}
@@ -472,10 +472,10 @@ XUIAutocompleter.propTypes = {
   /** Maps to the `size` property of the dropdown component. */
   dropdownSize: PropTypes.oneOf(Object.keys(fixedWidthDropdownSizes)),
 
-  /** Maps to the `closeOnSelect` property of the DropDownToggled component. */
+  /** Maps to the `closeOnSelect` property of the XUIDropDownToggled component. */
   closeOnSelect: PropTypes.bool,
 
-  /** Maps to the `closeOnTab` property of the DropDownToggled component. Set to false, if you've
+  /** Maps to the `closeOnTab` property of the XUIDropDownToggled component. Set to false, if you've
    * supplied a footer element with any links or interaction. */
   closeOnTab: PropTypes.bool,
 
@@ -494,7 +494,7 @@ XUIAutocompleter.propTypes = {
 
   /**
    * Setting to false will allow the dropdown's width to be set independent of the trigger width. <br>
-   * **Note:** *Setting this to true will override any size prop on DropDown.* <br>
+   * **Note:** *Setting this to true will override any size prop on XUIDropDown.* <br>
    * XUI design has also decided to keep a minimum width on the dropdown,
    * so dropdown may not match the width of narrow triggers.
    */