+++ src/react/components/select-box/SelectBox.d.ts
@@ -95,7 +95,7 @@ interface Props {
   /**
    * Setting to `false` will allow the dropdown's width to be set independent of the trigger width.
    *
-   * **Note:** *Setting this to true will override any size prop on `DropDown`.*
+   * **Note:** *Setting this to true will override any size prop on `XUIDropDown`.*
    *
    * XUI design has also decided to keep a minimum width on the dropdown, so dropdown may not match
    * the width of narrow triggers.
@@ -132,7 +132,7 @@ interface Props {
 
 export default class SelectBox extends React.Component<Props> {
   /**
-   * Check if the DropDown is open
+   * Check if the dropdown is open
    */
   isDropDownOpen(): boolean;
 }