+++ src/react/stories/components-in-components/dropdown-in-modal/index.js
@@ -3,7 +3,7 @@ import React from 'react';
 
 // Components we need to test with
 import XUIModal, { XUIModalBody } from '../../../modal';
-import DropDown, { DropDownToggled } from '../../../dropdown';
+import XUIDropDown, { XUIDropDownToggled } from '../../../dropdown';
 import Picklist, { Pickitem } from '../../../picklist';
 import XUIButton from '../../../button';
 
@@ -60,16 +60,16 @@ const toggledItems = [
 test.add(storyNames.dropDownInModal, () => {
   const trigger = <XUIButton hasCaret>Trigger Button</XUIButton>;
   const dropdown = (
-    <DropDown>
+    <XUIDropDown>
       <Picklist>{toggledItems}</Picklist>
-    </DropDown>
+    </XUIDropDown>
   );
 
   return (
     <XUIModal closeButtonLabel="Close" isOpen>
       <XUIModalBody>
         This is some Modal content.
-        <DropDownToggled dropdown={dropdown} trigger={trigger} />
+        <XUIDropDownToggled dropdown={dropdown} trigger={trigger} />
       </XUIModalBody>
     </XUIModal>
   );