+++ changes.md
@@ -37,6 +37,14 @@ Following classes have been **removed** because left space isn't needed for alig
 ## XUI React components
 
 - A number of React components have been converted from class-based to functional. You may need to make changes accordingly. Here’s [the list of affected components and more information about the change](https://docs.google.com/document/d/1x2vwW-cYZaX2hVmDk4rQBtoP-pp6Q2PNJbFY5-n7t78/edit).
+- We’ve renamed some XUI components to standardise the naming convention.
+  - Renamed components:
+    - `DropDown` → `XUIDropDown`
+    - `DropDownToggled` → `XUIDropDownToggled`
+    - `DropDownHeader` → `XUIDropDownHeader`
+    - `DropDownFooter` → `XUIDropDownFooter`
+    - `DropDownPanel` → `XUIDropDownPanel`
+    - `NestedDropDown` → `XUINestedDropDown`
 - `XUIStepper`
   - Updated spacing to match other components and 4px grid
 - `XUIButton`, `XUISecondaryButton`, `XUISplitButtonGroup`